+++ lib/state_pulse.dart
@@ -6,3 +6,5 @@ export 'src/state_pulse_builder.dart';
 export 'src/state_pulse_selector.dart';
 export 'src/state_pulse_listener.dart';
 export 'src/state_pulse_consumer.dart';
+export 'src/multi_state_pulse_provider.dart';
+export 'src/multi_state_pulse_listener.dart';